+++ src/data/Schedule.java
@@ -1,0 +1,4 @@
+package data;
+public class Schedule {
+	
+}
