+++ src/users/Admin.java
@@ -1,0 +1,5 @@
+package users;
+
+public class Admin {
+
+}
